// frontend/pharmacy-app/src/enums/nycbsPharmaOrder.ts
export enum NycbsPharmaOrderKeys {
  ActiveInsurance = "active_insurance",
  Identifier = "identifier",
  DrugName = "drug_name",
  DrugNameOncoEmr = "drug_name_onco_emr",
  PrimaryDiagnoses = "primary_diagnoses",
  PrimaryDiagnosesDescription = "primary_diagnoses_description",
  SecondaryDiagnoses = "secondary_diagnoses",
  SecondaryDiagnosesDescription = "secondary_diagnoses_description",
  PatientInsuranceState = "patient_insurance_state",
  PlanName = "plan_name",
  PatientRxBin = "patient_rx_bin",
  SftpRxBin = "sftp_rx_bin",
  PatientRxGroup = "patient_rx_group",
  PatientRxPcn = "patient_rx_pcn",
  PatientMemberId = "patient_member_id",
  FormName = "form_name",
  PatientFirstName = "patient_first_name",
  PatientLastName = "patient_last_name",
  PatientDob = "patient_dob",
  PatientGender = "patient_gender",
  PatientAddress = "patient_address",
  PatientPhoneNumber = "patient_phone_number",
  DrugQuantity = "drug_quantity",
  DrugQuantityQualifier = "drug_quantity_qualifier",
  DrugDaysSupply = "drug_days_supply",
  ProviderNpi = "provider_npi",
  ProviderFirstName = "provider_first_name",
  ProviderLastName = "provider_last_name",
  ProviderAddress = "provider_address",
  ProviderPhone = "provider_phone",
  ProviderFax = "provider_fax",
  ReviewType = "review_type",
  CreatedAt = "created_at",
  CmmResultKey = "cmm_result_key",
  PatientMrn = "patient_mrn",
  AssigneeId = "assignee_id",
  Status = "status",
  MedicationDrugName = "medication_drug_name",
  Reason = "reason",
  Comment = "comment",
  PaFormOptions = "pa_form_option",
  ErrorText = "error_text",
  FormNameInside = "form_name_inside",
  IsFormDirty = "is_form_dirty",
  PatientEligibilityCheckStatus = "patient_eligibility_check_status",
  IsDiagnosisCodeAvailableOnForm = "is_diagnosis_code_available_on_form",
  IsSecondaryDiagnosisCodeAvailableOnForm = "is_secondary_diagnosis_code_available_on_form",
  DrugFetchedFrom = "drug_fetched_from",
  PortalUsername = "portal_username",
  FormPickedFlag = "form_picked_flag",
  DrugConfidenceScore = "drug_confidence_score",
  DrugPickedThinking = "drug_picked_thinking",
  DiagnosisDetails = "diagnosis_details",
  Type = "type",
  PrimaryDiagnosesData = "primary_diagnoses_data",
  SecondaryDiagnosesData = "secondary_diagnoses_data",
  PrescriptionData = "prescription_data",
  DrugType = "drug_type",
  GenericBrand = "generic_brand",
  FormPicked = "form_picked",
}